Split main swift file
