+++ PidginMacOSItegration/Main.swift
@@ -27,176 +27,7 @@ extension String {
         }
     }
 }
-
-extension PidginConversation {
-    var latestImagePath : String? {
-        debug("Latest image path.")
-        var result : String?
-        if let entry = self.entry {
-            result = entry.withMemoryRebound(to: GtkIMHtml.self, capacity: 1) {
-                (html : UnsafeMutablePointer<GtkIMHtml>) -> String? in
-                if let images = html.pointee.im_images {
-                    let imageStruct = images.pointee.data.assumingMemoryBound(to: im_image_data.self)
-                    let image = html.pointee.funcs.pointee.image_get(imageStruct.pointee.id)
-                    if let imagePath = html.pointee.funcs.pointee.image_get_filename(image) {
-                        return String(cString: imagePath)
-                    }
-                }
-                return nil
-            }
-        }
-        debug("Latest image path: \(String(describing: result))")
-        log_all("macos", "Latest image path: \(String(describing: result))\n")
-        return result
-    }
-}
-
-extension PurpleMessageFlags {
-    var containsImages : Bool {
-        return self.rawValue & PURPLE_MESSAGE_IMAGES.rawValue > 0
-    }
-    
-    var sentFromRemote : Bool {
-        return self.rawValue & PURPLE_MESSAGE_REMOTE_SEND.rawValue > 0
-    }
-    
-    var sent : Bool {
-        return self.rawValue & PURPLE_MESSAGE_SEND.rawValue > 0
-    }
-    
-    var received : Bool {
-        return self.rawValue & PURPLE_MESSAGE_RECV.rawValue > 0
-    }
-    var system : Bool {
-        return self.rawValue & PURPLE_MESSAGE_SYSTEM.rawValue > 0
-    }
-    
-    var autoResponse : Bool {
-        return self.rawValue & PURPLE_MESSAGE_AUTO_RESP.rawValue > 0
-    }
-    
-    var activeOnly : Bool {
-        return self.rawValue & PURPLE_MESSAGE_ACTIVE_ONLY.rawValue > 0
-    }
-    
-    var nick : Bool {
-        return self.rawValue & PURPLE_MESSAGE_NICK.rawValue > 0
-    }
-    
-    var noLog : Bool {
-        return self.rawValue & PURPLE_MESSAGE_NO_LOG.rawValue > 0
-    }
-    
-    var whisper : Bool {
-        return self.rawValue & PURPLE_MESSAGE_WHISPER.rawValue > 0
-    }
-    var error : Bool { return self.rawValue & PURPLE_MESSAGE_ERROR.rawValue > 0 }
-    var delayed : Bool { return self.rawValue & PURPLE_MESSAGE_DELAYED.rawValue > 0 }
-    var raw : Bool { return self.rawValue & PURPLE_MESSAGE_RAW.rawValue > 0 }
-    var notify : Bool { return self.rawValue & PURPLE_MESSAGE_NOTIFY.rawValue > 0 } /**< Message is a notification */
-    var noLinkify : Bool { return self.rawValue & PURPLE_MESSAGE_NO_LINKIFY.rawValue > 0 }
-    var invisible : Bool { return self.rawValue & PURPLE_MESSAGE_INVISIBLE.rawValue > 0 } /**< Message should not be displayed */
-    
-    func shouldNotify() -> Bool {
-        debug("shouldNotify")
-        let toDisplay = (received || system) && !invisible && !sent
-        log_all("macos", "Notification should be displayed: \(toDisplay), flags: \(String(self.rawValue, radix: 2))\n")
-        return toDisplay
-    }
-}
-
-// FIXME: Rename to camelcase.
-func view_selection_func (selection : UnsafeMutablePointer<GtkTreeSelection>!, model : GtkTreeModel!, path : GtkTreePath!, path_currently_selected : gboolean, userdata : gpointer!) -> gboolean {
-    var iter = GtkTreeIter(stamp: 0, user_data: nil, user_data2: nil, user_data3: nil)
-    let result = gtk_tree_model_get_iter(model, &iter, path)
-    if (result == 1)
-    {
-        let index = Int(String(cString: gtk_tree_model_get_string_from_iter(model, &iter)))
-        if let i = index {
-            let n = Plugin.notificationSounds[i]
-            //        gtk_tree_model_get_value(model, &iter, 0, name)
-            //        gtk_tree_model_get(model, &iter, "Sound", name, -1)
-            if (path_currently_selected == FALSE)
-            {
-                log_all("macos", "\(n) is going to be selected.")
-//                let file = n.1.isEmpty ? n.0 : "\(n.1)/\(n.0)"
-                let file = n.0
-                purple_prefs_set_string(Plugin.Preferences.NotificationSound.path, file)
-                NSSound(named: file)?.play()
-            }
-            else
-            {
-                log_all("macos", "\(n) is going to be unselected.")
-            }
-        }
-    }
-    else {
-        log_critical("macos", "Selected nonexistent sound!")
-    }
-    return 1
-}
-
-func toPtr<T : AnyObject>(_ obj : T) -> UnsafeMutableRawPointer {
-    return UnsafeMutableRawPointer(Unmanaged.passUnretained(obj).toOpaque())
-}
-
-func fromPtr<T : AnyObject>(_ ptr : UnsafeMutableRawPointer) -> T {
-    debug("From ptr: \(String.init(describing: ptr))")
-    return Unmanaged<T>.fromOpaque(ptr).takeUnretainedValue()
-}
-
-func debug(_ message: String) {
-    #if DEBUG
-    fputs(message, __stderrp)
-    if(message.last != "\n")
-    {
-        fputs("\n", __stderrp)
-    }
-    log_all("macos", message)
-    #endif
-}
-
-func tryCast<T : AnyObject>(_ ptr: UnsafeMutableRawPointer?) -> T? {
-    if(ptr == nil)
-    {
-        debug("Received invalid pointer.\n")
-        return nil
-    }
-    return fromPtr(ptr!)
-}
-
-func forEachInList(_ list: UnsafeMutablePointer<GList>?, function: (UnsafeMutablePointer<GList>) -> ()) {
-    var element = list
-    while(element != nil) {
-        function(element!)
-        element = element!.pointee.next
-    }
-}
-enum GTypes : UInt {
-    case G_TYPE_INVALID = 0b0 // 0 << 2
-    case G_TYPE_NONE = 0b100 // 1 << 2
-    case G_TYPE_INTERFACE = 0b1000 // 2 << 2
-    case G_TYPE_CHAR = 0b1100 // 3 << 2
-    case G_TYPE_UCHAR = 0b10000 // 4 << 2
-    case G_TYPE_BOOLEAN = 0b10100 // 5 << 2
-    case G_TYPE_INT = 0b11000 // 6 << 2
-    case G_TYPE_UINT = 0b11100 // 7 << 2
-    case G_TYPE_LONG = 0b100000 // 8 << 2
-    case G_TYPE_ULONG = 0b100100 // 9 << 2
-    case G_TYPE_INT64 = 0b101000 // 10 << 2
-    case G_TYPE_UINT64 = 0b101100 // 11 << 2
-    case G_TYPE_ENUM = 0b110000 // 12 << 2
-    case G_TYPE_FLAGS = 0b110100 // 13 << 2
-    case G_TYPE_FLOAT = 0b111000 // 14 << 2
-    case G_TYPE_DOUBLE = 0b111100 // 15 << 2
-    case G_TYPE_STRING = 0b1000000 // 16 << 2
-    case G_TYPE_POINTER = 0b1000100 // 17 << 2
-    case G_TYPE_BOXED = 0b1001000 // 18 << 2
-    case G_TYPE_PARAM = 0b1001100 // 19 << 2
-    case G_TYPE_OBJECT = 0b1010000 // 20 << 2
-    case G_TYPE_VARIANT = 0b1010100 // 21 << 2
-}
-    
+   
 @objc class Plugin : NSObject, NSUserNotificationCenterDelegate, NSApplicationDelegate {
     var instance: UnsafeMutablePointer<PurplePlugin>?
     static let callbacks = ["receiving-im-msg", "receiving-chat-msg", /*"received-im-msg", "received-chat-msg",*/ "displayed-im-msg", "displayed-chat-msg"]
